MINOR: refactor: move {main,router,store}.js into web folder
