+++ .npmignore
@@ -3,6 +3,3 @@ src/**/__test__/
 
 # ignore design system app files
 src/web/
-src/main.js
-src/router.js
-src/store.js
