+++ src/web/main.js
@@ -2,7 +2,7 @@ import { createApp } from "vue";
 import VueClipboard from "vue3-clipboard";
 import router from "./router.js";
 
-import App from "./web/App.vue";
+import App from "./App.vue";
 
 const app = createApp(App);
 app.use(VueClipboard, {});
